requirements.txt
streamlit
openai
pandas
Pillow
PyMuPDF
